Struct prometheus_client::metrics::counter::Counter

pub struct Counter<N = u64, A = AtomicU64> { /* private fields */ }

Open Metrics Counter to measure discrete events.

Single monotonically increasing value metric.

Counter is generic over the actual data type tracking the Counter state as well as the data type used to interact with the Counter. Out of convenience the generic type parameters are set to use an AtomicU64 as a storage and u64 on the interface by default.

Examples

Using AtomicU64 as storage and u64 on the interface

let counter: Counter = Counter::default();
counter.inc();
let _value: u64 = counter.get();

Using AtomicU64 as storage and f64 on the interface

let counter = Counter::<f64, AtomicU64>::default();
counter.inc();
let _value: f64 = counter.get();

Implementations

impl<N, A: Atomic<N>> Counter<N, A>

pub fn inc(&self) -> N

Increase the Counter by 1, returning the previous value.

pub fn inc_by(&self, v: N) -> N

Increase the Counter by v, returning the previous value.

pub fn get(&self) -> N

Get the current value of the Counter.

pub fn inner(&self) -> &A

Exposes the inner atomic type of the Counter.

This should only be used for advanced use-cases which are not directly supported by the library.

The caller of this function has to uphold the property of an Open Metrics counter namely that the value is monotonically increasing, i.e. either stays the same or increases.