Module prometheus_client::metrics::exemplar

Module implementing an Open Metrics exemplars for counters and histograms.

See CounterWithExemplar and HistogramWithExemplars for details.

Structs

CounterWithExemplar

Open Metrics Counter with an Exemplar to both measure discrete events and track references to data outside of the metric set.

CounterWithExemplarInner

Exemplar

HistogramWithExemplars

Open Metrics Histogram to both measure distributions of discrete events. and track references to data outside of the metric set.

HistogramWithExemplarsInner