Struct prometheus_client::metrics::exemplar::CounterWithExemplar

pub struct CounterWithExemplar<S, N = u64, A = AtomicU64> { /* private fields */ }

Open Metrics Counter with an Exemplar to both measure discrete events and track references to data outside of the metric set.

let counter_with_exemplar = CounterWithExemplar::<Vec<(String, String)>>::default();
counter_with_exemplar.inc_by(1, Some(vec![("user_id".to_string(), "42".to_string())]));
let _value: (u64, _) = counter_with_exemplar.get();

You can also use exemplars with families. Just wrap the exemplar in a Family.

#[derive(Clone, Hash, PartialEq, Eq, EncodeLabelSet, Debug, Default)]
pub struct ResultLabel {
    pub result: String,
}

#[derive(Clone, Hash, PartialEq, Eq, EncodeLabelSet, Debug, Default)]
pub struct TraceLabel {
    pub trace_id: String,
}

let latency: Family<ResultLabel, CounterWithExemplar<TraceLabel>> = Family::default();

latency
    .get_or_create(&ResultLabel {
        result: "success".to_owned(),
    })
    .inc_by(
        1,
        Some(TraceLabel {
            trace_id: "3a2f90c9f80b894f".to_owned(),
        }),
    );

Implementations

impl<S, N: Clone, A: Atomic<N>> CounterWithExemplar<S, N, A>

pub fn inc_by(&self, v: N, label_set: Option<S>) -> N

Increase the CounterWithExemplar by v, updating the Exemplar if a label set is provided, returning the previous value.

pub fn get(&self) -> (N, MappedRwLockReadGuard<'_, Option<Exemplar<S, N>>>)

Get the current value of the CounterWithExemplar as well as its Exemplar if any.

pub fn inner(&self) -> MappedRwLockReadGuard<'_, A>

Exposes the inner atomic type of the CounterWithExemplar.

This should only be used for advanced use-cases which are not directly supported by the library.

The caller of this function has to uphold the property of an Open Metrics counter namely that the value is monotonically increasing, i.e. either stays the same or increases.