Module prometheus_client::registry

Metric registry implementation.

See Registry for details.

Structs

Descriptor

Registry

A metric registry.

RegistryIterator

Iterator iterating both the metrics registered directly with the registry as well as all metrics registered with sub-registries.

Enums

Unit

Metric units recommended by Open Metrics.